# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Driver and Delivery App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeliveryStatus =
  (typeof DeliveryStatus)[keyof typeof DeliveryStatus];

export const DeliveryStatus = {
  pending: "pending",
  in_progress: "in_progress",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type DeliveryPriority =
  (typeof DeliveryPriority)[keyof typeof DeliveryPriority];

export const DeliveryPriority = {
  low: "low",
  normal: "normal",
  urgent: "urgent",
} as const;

export interface Delivery {
  id: number;
  trackingNumber: string;
  customerName: string;
  customerPhone?: string;
  pickupAddress: string;
  deliveryAddress: string;
  status: DeliveryStatus;
  priority: DeliveryPriority;
  weight?: number;
  notes?: string;
  delivererId?: number;
  estimatedDeliveryTime?: string;
  createdAt: string;
  updatedAt?: string;
}

export type CreateDeliveryBodyPriority =
  (typeof CreateDeliveryBodyPriority)[keyof typeof CreateDeliveryBodyPriority];

export const CreateDeliveryBodyPriority = {
  low: "low",
  normal: "normal",
  urgent: "urgent",
} as const;

export interface CreateDeliveryBody {
  trackingNumber: string;
  customerName: string;
  customerPhone?: string;
  pickupAddress: string;
  deliveryAddress: string;
  priority: CreateDeliveryBodyPriority;
  weight?: number;
  notes?: string;
  delivererId?: number;
  estimatedDeliveryTime?: string;
}

export type UpdateDeliveryBodyStatus =
  (typeof UpdateDeliveryBodyStatus)[keyof typeof UpdateDeliveryBodyStatus];

export const UpdateDeliveryBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateDeliveryBody {
  status?: UpdateDeliveryBodyStatus;
  notes?: string;
  delivererId?: number;
}

export interface DeliveryStats {
  totalToday: number;
  completedToday: number;
  inProgress: number;
  pending: number;
  cancelled: number;
  earningsToday: number;
  earningsWeek: number;
  averageDeliveryTime: number;
}

export type DelivererVehicleType =
  (typeof DelivererVehicleType)[keyof typeof DelivererVehicleType];

export const DelivererVehicleType = {
  bicycle: "bicycle",
  motorcycle: "motorcycle",
  car: "car",
  van: "van",
} as const;

export type DelivererStatus =
  (typeof DelivererStatus)[keyof typeof DelivererStatus];

export const DelivererStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface Deliverer {
  id: number;
  name: string;
  phone: string;
  email?: string;
  vehicleType: DelivererVehicleType;
  status: DelivererStatus;
  zone?: string;
  totalDeliveries: number;
  rating: number;
  createdAt?: string;
}

export type CreateDelivererBodyVehicleType =
  (typeof CreateDelivererBodyVehicleType)[keyof typeof CreateDelivererBodyVehicleType];

export const CreateDelivererBodyVehicleType = {
  bicycle: "bicycle",
  motorcycle: "motorcycle",
  car: "car",
  van: "van",
} as const;

export interface CreateDelivererBody {
  name: string;
  phone: string;
  email?: string;
  vehicleType: CreateDelivererBodyVehicleType;
  zone?: string;
}

export type UpdateDelivererBodyStatus =
  (typeof UpdateDelivererBodyStatus)[keyof typeof UpdateDelivererBodyStatus];

export const UpdateDelivererBodyStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface UpdateDelivererBody {
  status?: UpdateDelivererBodyStatus;
  zone?: string;
}

export type TripStatus = (typeof TripStatus)[keyof typeof TripStatus];

export const TripStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Trip {
  id: number;
  passengerName: string;
  passengerPhone?: string;
  pickupAddress: string;
  dropoffAddress: string;
  status: TripStatus;
  fare: number;
  distance?: number;
  duration?: number;
  driverId?: number;
  scheduledAt?: string;
  startedAt?: string;
  completedAt?: string;
  createdAt: string;
}

export interface CreateTripBody {
  passengerName: string;
  passengerPhone?: string;
  pickupAddress: string;
  dropoffAddress: string;
  fare: number;
  distance?: number;
  driverId?: number;
  scheduledAt?: string;
}

export type UpdateTripBodyStatus =
  (typeof UpdateTripBodyStatus)[keyof typeof UpdateTripBodyStatus];

export const UpdateTripBodyStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateTripBody {
  status?: UpdateTripBodyStatus;
  startedAt?: string;
  completedAt?: string;
  fare?: number;
}

export interface TripStats {
  totalToday: number;
  completedToday: number;
  inProgress: number;
  cancelled: number;
  earningsToday: number;
  earningsWeek: number;
  totalKmToday: number;
  averageFare: number;
}

export type DriverStatus = (typeof DriverStatus)[keyof typeof DriverStatus];

export const DriverStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface Driver {
  id: number;
  name: string;
  phone: string;
  email?: string;
  licenseNumber: string;
  vehicleModel: string;
  vehiclePlate: string;
  status: DriverStatus;
  totalTrips: number;
  rating: number;
  createdAt?: string;
}

export interface CreateDriverBody {
  name: string;
  phone: string;
  email?: string;
  licenseNumber: string;
  vehicleModel: string;
  vehiclePlate: string;
}

export type UpdateDriverBodyStatus =
  (typeof UpdateDriverBodyStatus)[keyof typeof UpdateDriverBodyStatus];

export const UpdateDriverBodyStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface UpdateDriverBody {
  status?: UpdateDriverBodyStatus;
  vehicleModel?: string;
  vehiclePlate?: string;
}

export type DispatchResultPhase =
  (typeof DispatchResultPhase)[keyof typeof DispatchResultPhase];

export const DispatchResultPhase = {
  primary: "primary",
  cascade: "cascade",
} as const;

export interface DispatchResult {
  delivery: Delivery;
  assignedDelivererId?: number;
  assignedDelivererName?: string;
  phase: DispatchResultPhase;
  message: string;
}

export interface AcceptDeliveryBody {
  delivererId: number;
}

export interface ConfirmDeliveredBody {
  delivererId: number;
  proofNote?: string;
}

export interface PendingDispatch {
  hasPending: boolean;
  delivery?: Delivery;
  expiresAt?: string;
  secondsLeft?: number;
  phase?: string;
}

export type ListDeliveriesParams = {
  status?: ListDeliveriesStatus;
  delivererId?: number;
};

export type ListDeliveriesStatus =
  (typeof ListDeliveriesStatus)[keyof typeof ListDeliveriesStatus];

export const ListDeliveriesStatus = {
  pending: "pending",
  in_progress: "in_progress",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type GetMyPendingDispatchParams = {
  delivererId: number;
};

export type GetPendingDispatchParams = {
  delivererId: number;
};

export type GetDeliveryStatsParams = {
  delivererId?: number;
};

export type ListTripsParams = {
  status?: ListTripsStatus;
  driverId?: number;
};

export type ListTripsStatus =
  (typeof ListTripsStatus)[keyof typeof ListTripsStatus];

export const ListTripsStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetTripStatsParams = {
  driverId?: number;
};