# lib/api-zod/src/generated/types/confirmDeliveredBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Driver and Delivery App API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmDeliveredBody {
  delivererId: number;
  proofNote?: string;
}